Refactor to IPcamera implementing interfaces of functions it has.
