--- backend/src/main/java/com/benine/backend/camera/ipcameracontrol/MovingIPCamera.java
@@ -1,140 +1,0 @@
-package com.benine.backend.camera.ipcameracontrol;
-
-import com.benine.backend.camera.Camera;
-import com.benine.backend.camera.CameraConnectionException;
-import com.benine.backend.camera.MovingCamera;
-
-import java.text.DecimalFormat;
-import java.text.NumberFormat;
-
-/**
- * Concrete decorator of an IP Camera.
- * @author Bryan
- */
-public class MovingIPCamera implements MovingCamera {
-  
-  Camera camera;
-  
-  public MovingIPCamera(Camera camera) {
-    this.camera = camera;
-  }
-
-  /**
-   * Supported range:
-   * pan: -175 to 175 degrees.
-   * tilt: -30 to 210 degrees.
-   * pan speed: 1 to 30.
-   * tilt speed: 0 to 2.
-   * @param pan in degrees horizontal axis.
-   * @param tilt in degrees vertical axis.
-   * @param panSpeed integer to specify the speed of the pan movement.
-   * @param tiltSpeed integer to specify the speed of the tilt movement.
-   * @throws CameraConnectionException when command can not be completed.
-   */
-  @Override
-  public void moveTo(double pan, double tilt, int panSpeed, int tiltSpeed) 
-                                                                throws CameraConnectionException {
-    sendCommand("%23APS" + convertPanToHex(pan) + convertTiltToHex(tilt) 
-                    + convertPanSpeedtoHex(panSpeed) + convertTiltSpeed(tiltSpeed));
-  }
-  
-  /**
-   * Values must be between 1 and 99 otherwise they will be rounded.
-   * Hereby is 1 max speed to left or downward.
-   * 99 is max speed to right or upward.
-   * @param pan movement direction over horizontal axis.
-   * @param tilt movement direction over vertical axis.
-   * @throws CameraConnectionException when command can not be completed.
-   */
-  @Override
-  public void move(int pan, int tilt) throws CameraConnectionException {
-    pan = Math.max(1, pan);
-    pan = Math.min(99, pan);
-    tilt = Math.max(1, tilt);
-    tilt = Math.min(99, tilt);
-    NumberFormat formatter = new DecimalFormat("00");
-    sendCommand("%23PTS" + formatter.format(pan) + formatter.format(tilt));
-  }
-
-  @Override
-  public double[] getPosition() throws CameraConnectionException {
-    String res = sendCommand("%23APC");
-    if (res.substring(0, 3).equals("aPC")) {
-      return new double[]{convertPanToDouble(res.substring(3, 7)),
-                                  convertTiltToDouble(res.substring(7))};
-    } else {
-      throw new IpcameraConnectionException("Getting the position of the camera failed.");
-    }
-  }
-  
-  /**
-   * Convert the pan speed to hexadecimal form.
-   * And ensure it is in the range the camera supports.
-   * @param pan input pan speed.
-   * @return pan speed in hexadecimal form.
-   */
-  private String convertPanSpeedtoHex(int pan) {
-    pan = Math.min(29, pan);
-    pan = Math.max(0, pan);
-    return Integer.toHexString(pan);
-  }
-  
-  /**
-   * Convert the hexadecimal presentation of the tilt position to degrees.
-   * @param tilt position in hexadecimal.
-   * @return tilt position in degrees.
-   */
-  private double convertTiltToDouble(String tilt) {
-    return (int)(((Integer.valueOf(tilt, 16) - 7284) / 121.3541667) + 0.5) - 30;
-  }
- 
-  /**
-   * Convert the tilt position in a range the camera supports.
-   * @param tilt input position in degrees.
-   * @return tilt position in hexadecimal form.
-   */
-  private String convertTiltToHex(double tilt) {
-    tilt = Math.min(210, tilt);
-    tilt = Math.max(-30, tilt);
-    tilt = ((tilt + 30) * 121.3541667) + 7284;
-    return Integer.toHexString((int) (tilt + 0.5));
-  }
-  
-  /**
-   * Convert the tilt speed to be only 0, 1 or 2.
-   * @param tilt input speed.
-   * @return supported speed by the camera.
-   */
-  private int convertTiltSpeed(int tilt) {
-    tilt = Math.min(2, tilt);
-    tilt = Math.max(0, tilt);
-    return tilt;
-  }
-  
-  /**
-   * Converts hexadecimal representation to degrees.
-   * @param pan hexadecimal pan position.
-   * @return pan position in degrees.
-   */
-  private double convertPanToDouble(String pan) {
-    return (int)(((Integer.valueOf(pan, 16) - 11530) / 121.3628571) + 0.5) - 175;
-  }
-  
-  /**
-   * Convert the pan position in a range the camera supports.
-   * @param pan input position in degrees.
-   * @return pan position in hexadecimal form.
-   */
-  private String convertPanToHex(double pan) {
-    pan = Math.min(175, pan);
-    pan = Math.max(-175, pan);
-    pan = ((pan + 175) * 121.3628571) + 11530;
-
-    return Integer.toHexString((int) (pan + 0.5));
-  }
-
-  @Override
-  public String sendCommand(String cmd) throws CameraConnectionException {
-    return camera.sendCommand(cmd);
-  }
-}
